+++ src/models/position/batch.js
@@ -42,7 +42,7 @@ const _initKnex = async () => {
 }
 
 const startDbTransaction = async () => {
-  _initKnex()
+  await _initKnex()
   const trx = await knex.transaction()
   return trx
 }
@@ -68,8 +68,23 @@ const getLatestTransferStatesByTransferIdList = async (transfersIdList) => {
   }
 }
 
+const getParticipantCurrencyIds = async (trx, accountIds) => {
+  const participantCurrencies = await knex('participantCurrency')
+    .transacting(trx)
+    .whereIn('participantCurrencyId', accountIds)
+    .select('*')
+  return participantCurrencies
+}
+
+const getParticipantCurrencyIdsByParticipantIds = async (trx, participantIds) => {
+  const participantCurrencies = await knex('participantCurrency')
+    .transacting(trx)
+    .whereIn('participantId', participantIds)
+    .select('*')
+  return participantCurrencies
+}
+
 const getPositionsByAccountIdsForUpdate = async (trx, accountIds) => {
-  _initKnex()
   const participantPositions = await knex('participantPosition')
     .transacting(trx)
     .whereIn('participantCurrencyId', accountIds)
@@ -95,7 +110,6 @@ const getPositionsByAccountIdsNonTrx = async (accountIds) => {
 }
 
 const updateParticipantPosition = async (trx, participantPositionId, participantPositionValue, participantPositionReservedValue = null) => {
-  _initKnex()
   const optionalValues = {}
   if (participantPositionReservedValue !== null) {
     optionalValues.reservedValue = participantPositionReservedValue
@@ -114,5 +128,7 @@ module.exports = {
   getLatestTransferStatesByTransferIdList,
   getPositionsByAccountIdsForUpdate,
   getPositionsByAccountIdsNonTrx,
+  getParticipantCurrencyIds,
+  getParticipantCurrencyIdsByParticipantIds,
   updateParticipantPosition
 }